pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "nltk>=3.9.2",
    "pillow>=12.0.0",
    "qrcode>=8.2",
    "reportlab>=4.4.4",
    "scikit-learn>=1.7.2",
]
